requirements.txt
feedparser>=6.0
trafilatura>=1.9
requests>=2.31
pillow>=10.0
fastapi>=0.110
uvicorn[standard]>=0.29
jinja2>=3.1
python-multipart>=0.0.9
python-dotenv>=1.0
